+++ src/extension/Create Procedure/Code Action Creator/CodeActionCreatorAL0118.ts
@@ -6,6 +6,8 @@ import { CreateProcedureAL0118 } from '../Procedure Creator/CreateProcedureAL011
 import { CreateProcedure } from '../Procedure Creator/CreateProcedure';
 import { ALProcedure } from '../../Entities/alProcedure';
 import { CreateProcedureCommands } from '../CreateProcedureCommands';
+import { CreateProcedureAL0118IntegrationEvent } from '../Procedure Creator/CreateProcedureAL0118IntegrationEvent';
+import { CreateProcedureAL0118BusinessEvent } from '../Procedure Creator/CreateProcedureAL0118BusinessEvent';
 
 export class CodeActionCreatorAL0118 implements ICodeActionCreator {
     syntaxTree: SyntaxTree | undefined;
@@ -24,27 +26,38 @@ export class CodeActionCreatorAL0118 implements ICodeActionCreator {
     }
 
     async createCodeActions(): Promise<vscode.CodeAction[]> {
+        let codeActions: vscode.CodeAction[] = [];
         let createprocedureAL0118: CreateProcedureAL0118 = new CreateProcedureAL0118(this.document, this.diagnostic);
         let procedure: ALProcedure = await CreateProcedure.createProcedure(createprocedureAL0118);
-        let codeAction: vscode.CodeAction | undefined = await this.createCodeAction(this.document, this.diagnostic, procedure);
-        return codeAction ? [codeAction] : [];
+        let codeActionProcedure: vscode.CodeAction = await this.createCodeAction(procedure, 'Create Procedure ' + procedure.name, this.document, this.diagnostic);
+        codeActionProcedure.isPreferred = true;
+        
+        if (procedure.name.match(/^On[A-Z].*$/)) {
+            codeActionProcedure.isPreferred = false;
+
+            let createProcedureAL0118IntegrationEvent: CreateProcedureAL0118IntegrationEvent = new CreateProcedureAL0118IntegrationEvent(this.document, this.diagnostic);
+            let integrationEvent: ALProcedure = await CreateProcedure.createProcedure(createProcedureAL0118IntegrationEvent);
+            let codeActionIntegrationEvent: vscode.CodeAction = await this.createCodeAction(integrationEvent, 'Create IntegrationEvent Publisher ' + integrationEvent.name, this.document, this.diagnostic);
+            codeActionIntegrationEvent.isPreferred = true;
+            codeActions.push(codeActionIntegrationEvent);
+
+            let createProcedureAL0118BusinessEvent: CreateProcedureAL0118BusinessEvent = new CreateProcedureAL0118BusinessEvent(this.document, this.diagnostic);
+            let businessEvent: ALProcedure = await CreateProcedure.createProcedure(createProcedureAL0118BusinessEvent);
+            let codeActionBusinessEvent: vscode.CodeAction = await this.createCodeAction(businessEvent, 'Create BusinessEvent Publisher ' + businessEvent.name, this.document, this.diagnostic);
+            codeActions.push(codeActionBusinessEvent);
+        }
+
+        codeActions.push(codeActionProcedure);
+        return codeActions;
     }
 
-    private async createCodeAction(currentDocument: vscode.TextDocument, diagnostic: vscode.Diagnostic, procedureToCreate: ALProcedure): Promise<vscode.CodeAction | undefined> {
-        let codeActionToCreateProcedure: vscode.CodeAction =
-            await this.createFixToCreateProcedure(procedureToCreate, currentDocument, diagnostic);
-
-        return codeActionToCreateProcedure;
-    }
-
-    private async createFixToCreateProcedure(procedure: ALProcedure, document: vscode.TextDocument, diagnostic: vscode.Diagnostic): Promise<vscode.CodeAction> {
-        const codeAction = new vscode.CodeAction(`Create procedure ${procedure.name}`, vscode.CodeActionKind.QuickFix);
+    private async createCodeAction(procedure: ALProcedure, msg: string, document: vscode.TextDocument, diagnostic: vscode.Diagnostic): Promise<vscode.CodeAction> {
+        const codeAction = new vscode.CodeAction(msg, vscode.CodeActionKind.QuickFix);
         codeAction.command = {
             command: CreateProcedureCommands.createProcedureCommand,
             title: 'Create Procedure',
             arguments: [document, diagnostic, procedure]
         };
-        codeAction.isPreferred = true;
         return codeAction;
     }
 }