+++ src/extension/Create Procedure/Procedure Creator/CreateProcedureAL0118IntegrationEvent.ts
@@ -1,0 +1,18 @@
+import * as vscode from 'vscode';
+import { CreateProcedureAL0118 } from './CreateProcedureAL0118';
+import { ICreateProcedure } from './ICreateProcedure';
+
+export class CreateProcedureAL0118IntegrationEvent extends CreateProcedureAL0118 implements ICreateProcedure {
+    constructor(document: vscode.TextDocument, diagnostic: vscode.Diagnostic) {
+        super(document, diagnostic);
+    }
+    getMemberAttributes(): string[] {
+        return ['IntegrationEvent(false,false)'];
+    }
+    async getReturnType(): Promise<string | undefined> {
+        return undefined;
+    }
+    getJumpToCreatedProcedure(): boolean {
+        return false;
+    }
+}